+++ examples/hepshell_settings.py
@@ -1,0 +1,8 @@
+
+'''
+    List of modules which should be loaded as commands for hepshell
+'''
+COMMANDS = [
+    'hepshell.commands',
+    'mymodule.commands',
+]